refactor(racing) : 자동차 전진관련 메서드 racing->car로 이동
