+++ src/main/java/RacingGame/Car.java
@@ -1,5 +1,7 @@
 package RacingGame;
 
+import java.util.Random;
+
 public class Car {
     private int position;
 
@@ -13,5 +15,18 @@ public class Car {
 
     public void move() {
         this.position += 1;
+    }
+
+    public int getRandomNumber() {
+        Random random = new Random();
+        return random.nextInt(10);
+    }
+
+    public boolean isMoveCar(int randomNumber) {
+        if (randomNumber > 4) {
+            return true;
+        }
+
+        return false;
     }
 }
